+++ WebApi/Logic/Services/Interfaces/IAdminCrudService.cs
@@ -5,10 +5,10 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IAdminCrudService
     {
         Task<IEnumerable<Admin>> GetAllAdminsAsync();
-        Task<Admin?> GetAdminByIdAsync(int id);
+        Task<Admin?> GetAdminByIdAsync(long id);
         Task<Admin> AddAdminAsync(Admin admin);
         Task<Admin> UpdateAdminAsync(Admin admin);
-        Task DeleteAdminAsync(int id);
+        Task DeleteAdminAsync(long id);
         Task<Admin?> GetAdminByUsernameAsync(string username);
         Task<Admin> AddAdminFromRawPasswordAsync(string username, string password);
     }
