+++ WebApi/Logic/Services/Interfaces/IBalanceHistoryCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IBalanceHistoryCrudService
     {
         Task<IEnumerable<BalanceHistory>> GetAllHistoriesAsync();
-        Task<BalanceHistory?> GetHistoryByIdAsync(int id);
+        Task<BalanceHistory?> GetHistoryByIdAsync(long id);
         Task<BalanceHistory> AddHistoryAsync(BalanceHistory history);
         Task<BalanceHistory> UpdateHistoryAsync(BalanceHistory history);
-        Task DeleteHistoryAsync(int id);
+        Task DeleteHistoryAsync(long id);
     }
 }
