+++ WebApi/Logic/Services/Interfaces/IBalanceHistoryStatusCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IBalanceHistoryStatusCrudService
     {
         Task<IEnumerable<BalanceHistoryStatus>> GetAllStatusesAsync();
-        Task<BalanceHistoryStatus?> GetStatusByIdAsync(int id);
+        Task<BalanceHistoryStatus?> GetStatusByIdAsync(long id);
         Task<BalanceHistoryStatus> AddStatusAsync(BalanceHistoryStatus status);
         Task<BalanceHistoryStatus> UpdateStatusAsync(BalanceHistoryStatus status);
-        Task DeleteStatusAsync(int id);
+        Task DeleteStatusAsync(long id);
     }
 }
