+++ WebApi/Logic/Services/Interfaces/IBranchCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IBranchCrudService
     {
         Task<IEnumerable<Branch>> GetAllBranchesAsync();
-        Task<Branch?> GetBranchByIdAsync(int id);
+        Task<Branch?> GetBranchByIdAsync(long id);
         Task<Branch> AddBranchAsync(Branch branch);
         Task<Branch> UpdateBranchAsync(Branch branch);
-        Task DeleteBranchAsync(int id);
+        Task DeleteBranchAsync(long id);
     }
 }
