+++ WebApi/Logic/Services/Interfaces/IBranchMenuCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IBranchMenuCrudService
     {
         Task<IEnumerable<BranchMenu>> GetAllBranchMenusAsync();
-        Task<BranchMenu?> GetBranchMenuByIdAsync(int id);
+        Task<BranchMenu?> GetBranchMenuByIdAsync(long id);
         Task<BranchMenu> AddBranchMenuAsync(BranchMenu menu);
         Task<BranchMenu> UpdateBranchMenuAsync(BranchMenu menu);
-        Task DeleteBranchMenuAsync(int id);
+        Task DeleteBranchMenuAsync(long id);
     }
 }
