+++ WebApi/Logic/Services/Interfaces/ICategoryCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface ICategoryCrudService
     {
         Task<IEnumerable<Category>> GetAllCategoriesAsync();
-        Task<Category?> GetCategoryByIdAsync(int id);
+        Task<Category?> GetCategoryByIdAsync(long id);
         Task<Category> AddCategoryAsync(Category category);
         Task<Category> UpdateCategoryAsync(Category category);
-        Task DeleteCategoryAsync(int id);
+        Task DeleteCategoryAsync(long id);
     }
 }
