+++ WebApi/Logic/Services/Interfaces/IClientCrudService.cs
@@ -5,10 +5,10 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IClientCrudService
     {
         Task<IEnumerable<Client>> GetAllClientsAsync();
-        Task<Client?> GetClientByIdAsync(int id);
+        Task<Client?> GetClientByIdAsync(long id);
         Task<Client> AddClientAsync(Client client);
         Task<Client> UpdateClientAsync(Client client);
-        Task DeleteClientAsync(int id);
+        Task DeleteClientAsync(long id);
     }
 
 }
