+++ WebApi/Logic/Services/Interfaces/IEmployeeCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IEmployeeCrudService
     {
         Task<IEnumerable<Employee>> GetAllEmployeesAsync();
-        Task<Employee?> GetEmployeeByIdAsync(int id);
+        Task<Employee?> GetEmployeeByIdAsync(long id);
         Task<Employee> AddEmployeeAsync(Employee employee);
         Task<Employee> UpdateEmployeeAsync(Employee employee);
-        Task DeleteEmployeeAsync(int id);
+        Task DeleteEmployeeAsync(long id);
     }
 }
