+++ WebApi/Logic/Services/Interfaces/IEmployeeRoleCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IEmployeeRoleCrudService
     {
         Task<IEnumerable<EmployeeRole>> GetAllRolesAsync();
-        Task<EmployeeRole?> GetRoleByIdAsync(int id);
+        Task<EmployeeRole?> GetRoleByIdAsync(long id);
         Task<EmployeeRole> AddRoleAsync(EmployeeRole role);
         Task<EmployeeRole> UpdateRoleAsync(EmployeeRole role);
-        Task DeleteRoleAsync(int id);
+        Task DeleteRoleAsync(long id);
     }
 }
