+++ WebApi/Logic/Services/Interfaces/IFeedbackCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IFeedbackCrudService
     {
         Task<IEnumerable<Feedback>> GetAllFeedbacksAsync();
-        Task<Feedback?> GetFeedbackByIdAsync(int id);
+        Task<Feedback?> GetFeedbackByIdAsync(long id);
         Task<Feedback> AddFeedbackAsync(Feedback feedback);
         Task<Feedback> UpdateFeedbackAsync(Feedback feedback);
-        Task DeleteFeedbackAsync(int id);
+        Task DeleteFeedbackAsync(long id);
     }
 }
