+++ WebApi/Logic/Services/Interfaces/IMenuPresetCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IMenuPresetCrudService
     {
         Task<IEnumerable<MenuPreset>> GetAllMenuPresetsAsync();
-        Task<MenuPreset?> GetMenuPresetByIdAsync(int id);
+        Task<MenuPreset?> GetMenuPresetByIdAsync(long id);
         Task<MenuPreset> AddMenuPresetAsync(MenuPreset preset);
         Task<MenuPreset> UpdateMenuPresetAsync(MenuPreset preset);
-        Task DeleteMenuPresetAsync(int id);
+        Task DeleteMenuPresetAsync(long id);
     }
 }
