+++ WebApi/Logic/Services/Interfaces/IMenuPresetItemCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IMenuPresetItemCrudService
     {
         Task<IEnumerable<MenuPresetItem>> GetAllMenuPresetItemsAsync();
-        Task<MenuPresetItem?> GetMenuPresetItemByIdAsync(int id);
+        Task<MenuPresetItem?> GetMenuPresetItemByIdAsync(long id);
         Task<MenuPresetItem> AddMenuPresetItemAsync(MenuPresetItem item);
         Task<MenuPresetItem> UpdateMenuPresetItemAsync(MenuPresetItem item);
-        Task DeleteMenuPresetItemAsync(int id);
+        Task DeleteMenuPresetItemAsync(long id);
     }
 }
