+++ WebApi/Logic/Services/Interfaces/IOrderCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IOrderCrudService
     {
         Task<IEnumerable<Order>> GetAllOrdersAsync();
-        Task<Order?> GetOrderByIdAsync(int id);
+        Task<Order?> GetOrderByIdAsync(long id);
         Task<Order> AddOrderAsync(Order order);
         Task<Order> UpdateOrderAsync(Order order);
-        Task DeleteOrderAsync(int id);
+        Task DeleteOrderAsync(long id);
     }
 }
