+++ WebApi/Logic/Services/Interfaces/IOrderItemCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IOrderItemCrudService
     {
         Task<IEnumerable<OrderItem>> GetAllOrderItemsAsync();
-        Task<OrderItem?> GetOrderItemByIdAsync(int id);
+        Task<OrderItem?> GetOrderItemByIdAsync(long id);
         Task<OrderItem> AddOrderItemAsync(OrderItem item);
         Task<OrderItem> UpdateOrderItemAsync(OrderItem item);
-        Task DeleteOrderItemAsync(int id);
+        Task DeleteOrderItemAsync(long id);
     }
 }
