+++ WebApi/Logic/Services/Interfaces/IProductCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IProductCrudService
     {
         Task<IEnumerable<Product>> GetAllProductsAsync();
-        Task<Product?> GetProductByIdAsync(int id);
+        Task<Product?> GetProductByIdAsync(long id);
         Task<Product> AddProductAsync(Product product);
         Task<Product> UpdateProductAsync(Product product);
-        Task DeleteProductAsync(int id);
+        Task DeleteProductAsync(long id);
     }
 }
