+++ WebApi/Logic/Services/Interfaces/IRatingCrudService.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IRatingCrudService
     {
         Task<IEnumerable<Rating>> GetAllRatingsAsync();
-        Task<Rating?> GetRatingByIdAsync(int id);
+        Task<Rating?> GetRatingByIdAsync(long id);
         Task<Rating> AddRatingAsync(Rating rating);
         Task<Rating> UpdateRatingAsync(Rating rating);
-        Task DeleteRatingAsync(int id);
+        Task DeleteRatingAsync(long id);
     }
 }
