+++ WebApi/Logic/Services/Interfaces/IRepository.cs
@@ -5,9 +5,9 @@ namespace WebApi.Logic.Services.Interfaces
     public interface IRepository<T> where T : class
     {
         Task<IEnumerable<T>> GetAllAsync();
-        Task<T?> GetByIdAsync(int id);
+        Task<T?> GetByIdAsync(long id);
         Task<T> AddAsync(T entity);
         Task<T> UpdateAsync(T entity);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
     }
 }
