+++ WebApi/Logic/Services/RatingCrudService.cs
@@ -17,7 +17,7 @@ namespace WebApi.Logic.Services
             return await _repository.GetAllAsync();
         }
 
-        public async Task<Rating?> GetRatingByIdAsync(int id)
+        public async Task<Rating?> GetRatingByIdAsync(long id)
         {
             return await _repository.GetByIdAsync(id);
         }
@@ -32,7 +32,7 @@ namespace WebApi.Logic.Services
             return await _repository.UpdateAsync(rating);
         }
 
-        public async Task DeleteRatingAsync(int id)
+        public async Task DeleteRatingAsync(long id)
         {
             await _repository.DeleteAsync(id);
         }
